requirements.txt
torch
torchvision
timm
numpy
pandas
pyyaml
scikit-learn
tqdm
matplotlib
fastapi 
uvicorn 
pillow
python-multipart